+++ boot/core/src/boot/core.clj
@@ -964,3 +964,20 @@
   "This function is the same as `by-re` but negated."
   [res files]
   (by-re res files true))
+
+;; General utilities ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(defn launch-nrepl
+  "Launches an nREPL server in a pod. See the repl task for options."
+  [& {:keys [pod] :as opts}]
+  (require 'boot.repl)
+  (let [mw   @@(resolve 'boot.repl/*default-middleware*)
+        deps @@(resolve 'boot.repl/*default-dependencies*)
+        opts (-> (dissoc opts :pod)
+                 (update-in [:init-ns] #(or % (when pod 'boot.pod)))
+                 (assoc :default-middleware mw :default-dependencies deps))]
+    (if (or (not pod) (= pod "core"))
+      (@(resolve 'boot.repl/launch-nrepl) opts)
+      (pod/with-eval-in (pod/get-pods pod true)
+        (require 'boot.repl)
+        (boot.repl/launch-nrepl '~opts)))))